Move createComponent util to the `ui` folder
