+++ packages/components/src/ui/utils/index.js
@@ -1,0 +1,1 @@
+export { createComponent } from './create-component';
