+++ packages/components/src/utils/index.js
@@ -1,3 +1,2 @@
 export * from './hooks';
 export * from './style-mixins';
-export { createComponent } from './create-component';
